autonet-kafka/DB_Server/requirements.txt
numpy
fastapi
Pillow
aiofiles
pytest
pytest-cov
pytest-asyncio
uvicorn
requests
python-multipart
